requirements.txt
discord.py
python-dotenv.py
matplotlib.py